requirements.txt
cython>=0.2
ujson~=1.35
tornado>=6.0,<7.0
msgpack~=0.6
pyzmq>=18.1.1

python-snappy~=0.5

arrow~=0.15

csirtg-dt==0.1
csirtg-re==0.1
csirtg-peers==0.3
csirtg-spamhaus==0.1

csirtg-indicator>=3.0b1,<4.0

csirtgsdk==1.1.5